TempFile should fail if dir is valid file path and O_TMPFILE is missing.
By passing O_DIRECTORY|O_RDWR to open(2) it is possible to ensure that,
open(2) will return an error if dir is the valid path to a file and the
kernel does not support O_TMPFILE.

Previously, given the above circumstances, open(2) would return a valid
file descriptor for the file at the given path. This compromised the
safety of TempFile on pre-O_TMPFILE kernels.